Fix: useQuery 부분 -커스텀훅 적용
