+++ package.json
@@ -23,6 +23,7 @@
     "react-youtube": "^10.1.0",
     "sass": "^1.63.6",
     "timeago.js": "^4.0.2",
+    "uuid": "^9.0.0",
     "vite-plugin-mkcert": "^1.16.0",
     "vite-plugin-sass-dts": "^1.3.6"
   },
@@ -31,6 +32,7 @@
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
     "@types/react-js-pagination": "^3.0.4",
+    "@types/uuid": "^9.0.2",
     "@typescript-eslint/eslint-plugin": "^5.61.0",
     "@typescript-eslint/parser": "^5.61.0",
     "@vitejs/plugin-react-swc": "^3.3.2",
