+++ src/Routes/PretendBuy/Pretend-Buy.tsx
@@ -6,6 +6,7 @@ export default function PretendBuy() {
   const [whatYear, setWhatYear] = useState(new Date().getFullYear());
   const [whatMonth, setWhatMonth] = useState(new Date().getMonth() + 1);
   const [today, setToday] = useState(new Date().getDate());
+
   let todayDate: string = `${whatYear}-${String(whatMonth).padStart(2, "0")}-${String(today).padStart(2, "0")}`;
   return (
     <>
